+++ src/main/java/com/livraria/controller/BookController.java
@@ -19,28 +19,31 @@ public class BookController {
 		this.bookView = bookView;
 	}
 
-	public void startBookController() {
+	public void startController() {
 		// listener para o botão de adicionar livros
 		bookView.addActionListener(new ActionListener() {
 			@Override
 			public void actionPerformed(ActionEvent e) {
-				String title = bookView.getTitleInput();
-				String isbn = bookView.getIsbnInput();
-				int publisher_id = Integer.parseInt(bookView.getPublisherInput());
-				double price = Double.parseDouble(bookView.getPriceInput());
-				
-				Books book = new Books(title, isbn, publisher_id, price);
+				if(verifyEmptyString(bookView.getTitleInput(), bookView.getIsbnInput(),
+						bookView.getPriceInput(), bookView.getPublisherInput())) {
+					String title = bookView.getTitleInput();
+					String isbn = bookView.getIsbnInput();
+					int publisherId = Integer.parseInt(bookView.getPublisherInput());
+					double price = Double.parseDouble(bookView.getPriceInput());
 
-				try {
-					boolean res = bookDao.addBooks(book);
-					if(res) {
-						bookView.showMessage("Livro adicionado com sucesso");
-					} else {
+					Books book = new Books(title, isbn, publisherId, price);
+
+					try {
+						boolean res = bookDao.addBooks(book);
+						if (res) {
+							bookView.showMessage("Livro adicionado com sucesso");
+						} else {
+							bookView.showMessage("Não foi possível adicionar o livro");
+						}
+					} catch (SQLIntegrityConstraintViolationException e1) {
 						bookView.showMessage("Não foi possível adicionar o livro");
+						e1.printStackTrace();
 					}
-				} catch (SQLIntegrityConstraintViolationException e1) {
-					bookView.showMessage("Não foi possível adicionar o livro");
-					e1.printStackTrace();
 				}
 			}
 		});
@@ -48,45 +51,64 @@ public class BookController {
 		bookView.delActionListener(new ActionListener() {
 			@Override
 			public void actionPerformed(ActionEvent e) {
-				String isbn = bookView.getIsbnInput();
-				System.out.println(isbn);
-				try {
-					boolean res1 = bookDao.deleteBooksAuthors(isbn);
-					boolean res2 = bookDao.deleteBooks(isbn);
-					if(res2) {
-						bookView.showMessage("Livro excluído com sucesso");
-					} else {
-						bookView.showMessage("Não foi possível excluir o livro");
+				if(verifyEmptyString(bookView.getIsbnInput())) {
+					String isbn = bookView.getIsbnInput();
+					System.out.println(isbn);
+					try {
+						boolean res1 = bookDao.deleteBooksAuthors(isbn);
+						boolean res2 = bookDao.deleteBooks(isbn);
+						if (res2) {
+							bookView.showMessage("Livro excluído com sucesso");
+						} else {
+							bookView.showMessage("Não foi possível excluir o livro");
+						}
+					} catch (SQLIntegrityConstraintViolationException e1) {
+						e1.printStackTrace();
 					}
-				} catch (SQLIntegrityConstraintViolationException e1) {
-					e1.printStackTrace();
 				}
 			}
 		});
-		// listener para o botão de modificar livros
+		// Listener para o botão de modificar livros
 		bookView.modifyActionListener(new ActionListener() {
 			@Override
 			public void actionPerformed(ActionEvent e) {
 				System.out.println("Botão funcionando");
+				if(verifyEmptyString(bookView.getTitleInput(), bookView.getIsbnInput(),
+						bookView.getPriceInput(), bookView.getPublisherInput())) {
+
+				}
 			}
 		});
 		// listener para o botão de pesquisar livros
 		bookView.searchActionListener(new ActionListener() {
 			@Override
 			public void actionPerformed(ActionEvent e) {
-				String title = bookView.getTitleInput();
-				List<Books> searchBookList = bookDao.searchBooksTitle(title);
-				DefaultListModel listModel = new DefaultListModel();
-				String[] columnNames = {"Título", "Editora", "Preço"};
-				DefaultTableModel model = new DefaultTableModel(columnNames, 0);
-				// guarda as instâncias de livros em listModel e adiciona as instâncias na tabela
-				for(Books book : searchBookList) {
-					listModel.addElement(book);
-					model.addRow(new Object[]{book.getTitle(), book.getPublisherId(), book.getPrice()});
+				if(verifyEmptyString(bookView.getTitleInput())) {
+					String title = bookView.getTitleInput();
+					List<Books> searchBookList = bookDao.searchBooksTitle(title);
+					DefaultListModel listModel = new DefaultListModel();
+					// Define as colunas da tabela
+					String[] columnNames = {"Título", "Editora", "Preço"};
+					DefaultTableModel model = new DefaultTableModel(columnNames, 0);
+					// guarda as instâncias de livros em listModel e adiciona as instâncias na tabela
+					for (Books book : searchBookList) {
+						listModel.addElement(book);
+						model.addRow(new Object[]{book.getTitle(), book.getPublisherId(), book.getPrice()});
+					}
+					bookView.showSearchResult(model);
 				}
-				bookView.showSearchResult(model);
 			}
 		});
 	}
-
+	// Verifica se o input não possui campos vazios
+	private boolean verifyEmptyString(String... inputs) {
+		for(int i = 0; i < inputs.length; i++) {
+			if(inputs[i].equals("")) {
+				bookView.showMessage("Preencha todos os campos");
+				return false;
+			}
+		}
+		System.out.println("No empty strings");
+		return true;
+	}
 }
